+++ src/vdxf/classes/index.ts
@@ -45,6 +45,8 @@ export {
 } from './provisioning/ProvisioningResult'
 export { Context } from './Context'
 export { DataCategory } from './PersonalProfile'
-export * from './attestation/InformationRequest'
+export * from './requestobjects/InformationRequest'
+export * from './requestobjects/LoginRequestDetails'
+export * from './requestobjects/ProvisioningIdentity'
 export * from './attestation/AttestationDetails'
 export * from './endorsement/Endorsement'
