+++ src/vdxf/classes/requestobjects/InformationRequest.ts
